+++ packages/iov-ethereum/src/serialization.spec.ts
@@ -1,8 +1,10 @@
 import {
   Address,
   Algorithm,
+  Amount,
   ChainId,
   Nonce,
+  PublicIdentity,
   PublicKeyBytes,
   SendTransaction,
   SignatureBytes,
@@ -98,6 +100,99 @@ describe("Serialization", () => {
         fromHex(
           "f8b7808504a817c8008252089443aa18faae961c23715735682dc75662d90f4dde8901158e460913d00000b887546865206e696365206d656d6f20492061747461636820746f2074686174206d6f6e657920666f72207468652077686f6c6520776f726c6420746f20726561642c20416e642063616e20656e636f6465206173206d756368206461746120617320796f752077616e742c20616e6420756e69636f64652073796d626f6c73206c696b6520e29da48216918080",
         ),
+      );
+    });
+
+    it("throws for unset gas price/limit", () => {
+      const creator: PublicIdentity = {
+        chainId: "ethereum-eip155-5777" as ChainId,
+        pubkey: {
+          algo: Algorithm.Secp256k1,
+          data: fromHex(
+            "044bc2a31265153f07e70e0bab08724e6b85e217f8cd628ceb62974247bb493382ce28cab79ad7119ee1ad3ebcdb98a16805211530ecc6cfefa1b88e6dff99232a",
+          ) as PublicKeyBytes,
+        },
+      };
+      const amount: Amount = {
+        quantity: "20000000000000000000",
+        fractionalDigits: 18,
+        tokenTicker: "ETH" as TokenTicker,
+      };
+      const gasPrice: Amount = {
+        quantity: "20000000000",
+        fractionalDigits: 18,
+        tokenTicker: "ETH" as TokenTicker,
+      };
+      const gasLimit: Amount = {
+        quantity: "21000",
+        fractionalDigits: 18,
+        tokenTicker: "ETH" as TokenTicker,
+      };
+      const nonce = new Int53(0) as Nonce;
+
+      // gasPrice unset
+      {
+        const tx: SendTransaction = {
+          kind: "bcp/send",
+          creator: creator,
+          amount: amount,
+          gasPrice: undefined,
+          gasLimit: gasLimit,
+          recipient: "0x43aa18FAAE961c23715735682dC75662d90F4DDe" as Address,
+        };
+        expect(() => serializeUnsignedTransaction(tx, nonce)).toThrowError(/gasPrice must be set/i);
+      }
+
+      // gasLimit unset
+      {
+        const tx: SendTransaction = {
+          kind: "bcp/send",
+          creator: creator,
+          amount: amount,
+          gasPrice: gasPrice,
+          gasLimit: undefined,
+          recipient: "0x43aa18FAAE961c23715735682dC75662d90F4DDe" as Address,
+        };
+        expect(() => serializeUnsignedTransaction(tx, nonce)).toThrowError(/gasLimit must be set/i);
+      }
+    });
+
+    it("throws for negative nonce", () => {
+      const creator: PublicIdentity = {
+        chainId: "ethereum-eip155-5777" as ChainId,
+        pubkey: {
+          algo: Algorithm.Secp256k1,
+          data: fromHex(
+            "044bc2a31265153f07e70e0bab08724e6b85e217f8cd628ceb62974247bb493382ce28cab79ad7119ee1ad3ebcdb98a16805211530ecc6cfefa1b88e6dff99232a",
+          ) as PublicKeyBytes,
+        },
+      };
+      const amount: Amount = {
+        quantity: "20000000000000000000",
+        fractionalDigits: 18,
+        tokenTicker: "ETH" as TokenTicker,
+      };
+      const gasPrice: Amount = {
+        quantity: "20000000000",
+        fractionalDigits: 18,
+        tokenTicker: "ETH" as TokenTicker,
+      };
+      const gasLimit: Amount = {
+        quantity: "21000",
+        fractionalDigits: 18,
+        tokenTicker: "ETH" as TokenTicker,
+      };
+
+      const tx: SendTransaction = {
+        kind: "bcp/send",
+        creator: creator,
+        amount: amount,
+        gasPrice: gasPrice,
+        gasLimit: gasLimit,
+        recipient: "0x43aa18FAAE961c23715735682dC75662d90F4DDe" as Address,
+      };
+      expect(() => serializeUnsignedTransaction(tx, new Int53(-1) as Nonce)).toThrowError(
+        /not a unsigned safe integer/i,
       );
     });
   });