+++ packages/iov-ethereum/src/utils.ts
@@ -28,13 +28,11 @@ export function decodeHexQuantityNonce(hexString: string): Nonce {
 }
 
 export function encodeQuantity(value: number): string {
-  if (Number.isNaN(value) || !/^[0-9]+$/.test(String(value))) {
-    throw new Error("Input is not a valid number");
-  }
   try {
-    return "0x" + new BN(value).toString(16);
+    const checkedValue = new Uint53(value);
+    return "0x" + new BN(checkedValue.toNumber()).toString(16);
   } catch {
-    throw new Error("Input is not a safe integer");
+    throw new Error("Input is not a unsigned safe integer");
   }
 }
 
@@ -45,11 +43,17 @@ export function encodeQuantityString(value: string): string {
   return "0x" + new BN(value).toString(16);
 }
 
-export function hexPadToEven(hex: string): string {
-  if (hex.length % 2 !== 0) {
-    return "0" + hex.replace("0x", "");
+/**
+ * Takes a hex representation optionally prefixed with 0x and returns a normalized
+ * representation: unprefixed, padded to even characters count, lower case.
+ */
+export function normalizeHex(input: string): string {
+  const unprefixedLower = input.replace("0x", "").toLowerCase();
+  if (unprefixedLower.length % 2 !== 0) {
+    return "0" + unprefixedLower;
+  } else {
+    return unprefixedLower;
   }
-  return hex.replace("0x", "");
 }
 
 export function toBcpChainId(numericChainId: number): ChainId {