+++ packages/iov-keycontrol/src/hdpaths.spec.ts
@@ -44,6 +44,25 @@ describe("HdPaths", () => {
     ]);
   });
 
+  it("has working Ethereum implementation", () => {
+    // m/44'/60'/0'/0/0
+    expect(HdPaths.ethereum(0)).toEqual([
+      Slip10RawIndex.hardened(44),
+      Slip10RawIndex.hardened(60),
+      Slip10RawIndex.hardened(0),
+      Slip10RawIndex.normal(0),
+      Slip10RawIndex.normal(0),
+    ]);
+    // m/44'/60'/0'/0/123
+    expect(HdPaths.ethereum(123)).toEqual([
+      Slip10RawIndex.hardened(44),
+      Slip10RawIndex.hardened(60),
+      Slip10RawIndex.hardened(0),
+      Slip10RawIndex.normal(0),
+      Slip10RawIndex.normal(123),
+    ]);
+  });
+
   it("has working MetaMask 'HD Key Tree' implementation", () => {
     // m/44'/60'/0'/0/0
     expect(HdPaths.metamaskHdKeyTree(0)).toEqual([