+++ packages/iov-keycontrol/types/hdpaths.d.ts
@@ -40,10 +40,18 @@ export declare class HdPaths {
      */
     static iov(account: number): ReadonlyArray<Slip10RawIndex>;
     /**
-     * The default MetaMask derivation path
+     * The default Ethereum derivation path
+     *
+     * This is compatible to MetaMask and Trezor.
      *
      * What MetaMask calls the account is a BIP44 address index: m/44'/60'/0'/0/<account>
      * (see https://github.com/MetaMask/eth-hd-keyring/blob/018a11a3a2/index.js#L8)
+     */
+    static ethereum(account: number): ReadonlyArray<Slip10RawIndex>;
+    /**
+     * The default MetaMask derivation path
+     *
+     * @deprecated use HdPaths.ethereum
      */
     static metamaskHdKeyTree(account: number): ReadonlyArray<Slip10RawIndex>;
 }